PythonBackned/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "keycloak-auth"
version = "0.1.0"
description = "Reusable Keycloak JWT validation for Python"
requires-python = ">=3.11"
dependencies = [
    "PyJWT[crypto]>=2.8",
    "httpx>=0.25",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
fastapi = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
]
dev = [
    "keycloak-auth[fastapi]",
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.25",
    "cryptography>=41.0",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/keycloak_auth"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
